+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # MiniReelinator
 
 * *[v1.29.0.rc1]*
+  * Show human readable interest labels in the campaign admin tab
   * [FEATURE]: Add Campaign Stats tab in manager area
 
 ## v1.28.1 (December 9, 2015)